Moved tests around into a better directory structure. Also updated tests to use a MockChatRepository instead of starting a Spring server every time we run the unit tests..
